Fixed useInsertionEffect Error
Took a bit, since there's very little on this bug and the docs weren't much help, but I tracked it down to the tRPC instance, specifically the SSR.

Appearently Tamagui uses useInsertionEffect in their compiled code and can't be removed.

I'm not too familiar with tRPC so if you have any ideas for a better solution that would be great!